+++ packages/connectivity/src/scp-cf/destination/destination-accessor-proxy-configuration.spec.ts
@@ -29,7 +29,7 @@ import {
 import { Protocol } from '../protocol';
 import { getDestination } from './destination-accessor';
 import { parseDestination } from './destination';
-import * as ProxyUtil from './proxy-util';
+import * as ProxyUtil from './http-proxy-util';
 import { alwaysProvider } from './destination-selection-strategies';
 import { Destination } from './destination-service-types';
 import { destinationCache } from './destination-cache';
