+++ packages/connectivity/src/scp-cf/destination/destination-from-registration.ts
@@ -13,7 +13,7 @@ import {
   addProxyConfigurationInternet,
   ProxyStrategy,
   proxyStrategy
-} from './proxy-util';
+} from './http-proxy-util';
 
 const logger = createLogger({
   package: 'connectivity',
