+++ packages/connectivity/src/scp-cf/destination/index.ts
@@ -10,5 +10,5 @@ export * from './destination-service-cache';
 export * from './destination-service-types';
 export * from './destination-service';
 export * from './destination';
-export * from './proxy-util';
+export * from './http-proxy-util';
 export * from './service-binding-to-destination';
